trap cleanup SIGNIT

+++ start_arm_control.sh
@@ -6,13 +6,17 @@ current_dir=$(pwd)
 
 trap cleanup SIGINT
 
+# Setup ROS log dir
+export ROS_LOG_DIR=$current_dir/.ros/log
+mkdir -p $ROS_LOG_DIR
+
 # Function to handle cleanup on Ctrl+C
 cleanup() {
   echo "Stopping all ROS nodes..."
   # Kill background processes
   pkill -P $$
-  kill -9 $INV_KIN_PID
-  kill -9 $NODE_ARM_PID
+  [ -n "$INV_KIN_PID" ] && kill -9 $INV_KIN_PID
+  [ -n "$NODE_ARM_PID" ] && kill -9 $NODE_ARM_PID
   # Stop ROS2 daemon (optional)
   ros2 daemon stop
   echo "All nodes stopped."
